slx150x2, hw: added demo design containing 2 FPGAs
The added design contains two XC6SLX150 in the FGG484 package. The design
demonstrates the FPGA side of the system, the power supply and the interface
are missing here. The FPGAs can be booted by JTAG and SPI and can communicate
through SPI afterwards. All pins are connected and the design is fully routed.