Added supply layers, various
Defined layers 2 and 15 as supply layers, removing them from the routing.
This required merging VCCINT0 and VCCINT1 again. Additionally, it now was
cheap to add a net NC for all unconnected pins of the FPGA.

All big components are on the mil 25 grid, just the small caps on the back
of the FPGAs are on a local 0.5mm grid around their FPGA.

Unfolded the busses. Now they need more space but should be "safe". Moved
around the termination resistors.