Added script to update BOM database
The script updateavail.sh (POSIX systems only) queries the distributors webpage
for each part in the *.tsv file and adds columns for number of on-stock parts
and prices for different price-breaks. The project.tsv file has been
reformatted to match the expected input of this script.

TODO: Currently supports only DigiKey as a supplier. Need to add Avnet support.